dsa: widen index-pool gather kernel offsets to int64
The DSA index-pool gather kernels compute source addresses as
page_index * buf_numel_per_page with page_index loaded from an int32
page table. The product overflows int32 once the pool exceeds 2^31
elements (page 254,203 with 8,448-element pages), producing a wild
address and an illegal memory access at runtime.

Same class of bug as #30859; widen the _GetK/_GetS/_GetKAndS triton
kernels used by every DSA index-K/S gather on the non-aiter path.

+++ python/sglang/kernels/ops/attention/dsa/index_buf_accessor.py
@@ -467,7 +467,7 @@ def _get_k_triton_kernel(
     token_offset_in_page = token_id % page_size
 
     # Load the page index from page_indices
-    page_index = tl.load(page_indices_ptr + page_idx)
+    page_index = tl.load(page_indices_ptr + page_idx).to(tl.int64)
 
     # Calculate source offset in buf
     # buf[page_index, token_offset_in_page * index_head_dim : ...]
@@ -544,7 +544,7 @@ def _get_s_triton_kernel(
     token_offset_in_page = token_id % page_size
 
     # Load the page index from page_indices
-    page_index = tl.load(page_indices_ptr + page_idx)
+    page_index = tl.load(page_indices_ptr + page_idx).to(tl.int64)
 
     # Calculate source offset in buf
     # Scales are stored after K data: page_size * index_head_dim offset
@@ -677,7 +677,7 @@ def _get_k_and_s_triton_kernel(
     page_index = tl.load(
         page_indices_ptr + page_idx + page_indices_base,
         mask=token_valid_mask & page_idx_valid_mask,
-    )
+    ).to(tl.int64)
 
     # ===== Load K data =====
     # The address calculation logic for K: page_index * total number of elements in a single page + K offset of the token within the page.
